requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
langchain-ollama>=0.1.0
pymongo>=4.6.0
requests>=2.31.0
pyyaml>=6.0
pydantic>=2.0.0
networkx>=3.0
vaderSentiment>=3.3.2
rich>=13.0.0
ollama>=0.1.7